Renamed SubmitFormButton to SubmitButton. Renamed types.ts to FormFieldsTypes.ts.
